add set -e and conditionally apply expected failure cases

+++ install.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 
+set -e
+
 # Install 1Password & 1Password CLI
 sudo rpm --import https://downloads.1password.com/linux/keys/1password.asc
 sudo sh -c 'echo -e "[1password]\nname=1Password Stable Channel\nbaseurl=https://downloads.1password.com/linux/rpm/stable/\$basearch\nenabled=1\ngpgcheck=1\nrepo_gpgcheck=1\ngpgkey=\"https://downloads.1password.com/linux/keys/1password.asc\"" > /etc/yum.repos.d/1password.repo'
@@ -16,17 +18,17 @@ op read op://personal/lpvmy5zudeushql4arboiunyhm/'public key' > ~/.ssh/id_ed2551
 chmod 700 ~/.ssh/id_ed25519
 
 # Install Yazi
-sudo dnf copr enable lihaohong/yazi
+sudo dnf copr enable -y lihaohong/yazi
 sudo dnf install -y zsh neovim fzf chromium crudini yazi
 sudo chsh -s $(which zsh) $USER
 
 # Clone dotfiles repo
-[ ! -f ~$HOME/.dotfiles ] && git clone --bare git@github.com:epictris/.dotfiles $HOME/.dotfiles 2>/dev/null
+[ ! -d $HOME/.dotfiles ] && git clone --bare git@github.com:epictris/.dotfiles $HOME/.dotfiles
 /usr/bin/git --git-dir=$HOME/.dotfiles/ --work-tree=$HOME pull origin main
 /usr/bin/git --git-dir=$HOME/.dotfiles/ --work-tree=$HOME submodule update --init --recursive --remote --merge
 
 # Install minimal SDDM theme
-sudo rm -r /usr/share/sddm/themes/minimal
+[ ! -d /usr/share/sddm/themes/minimal] && sudo rm -r /usr/share/sddm/themes/minimal
 sudo cp -r $HOME/.config/sddm-theme/minimal /usr/share/sddm/themes
 sudo cp $HOME/.config/background.jpg /usr/share/backgrounds/background.jpg
 [ ! -f /etc/sddm.conf.backup ] && sudo cp /etc/sddm.conf /etc/sddm.conf.backup
